update integer to string -> financeAPI

+++ controle-de-despesas/src/main/java/br/com/finance/cdd/service/AplicationServices.java
@@ -71,7 +71,7 @@ public class AplicationServices {
 
 	// Otimizaação de código
 	private Aplication findByIdApp(Long id) {
-		return appRepository.findById(id)
+		return appRepository.findById(id.longValue())
 				.orElseThrow(() -> new ResourceNotFoundException("Pay Not Found By ID: " + id));
 	}
